refactor: sort, category 별도 훅으로 추출
